+++ NEWS.md
@@ -12,6 +12,7 @@
 - updated `theme_schola()`
   - no margins by default (argument `margin` supersedes `margin_side` and `margin_bottom`)
   - axis titles on by default (+ `axis.title` renamed to `axis_title`)
+- figure cropping with `ghostscript` and `pdfcrop` is defunct, so those are no longer needed (see above)
 - new Google Drive handling -- URL is set in `.Rprofile` as a hidden object that can be accessed by `gd_get_proj()`
 
 ## Changes